+++ examples/error-handling/src/yoga.ts
@@ -1,0 +1,63 @@
+import { createYoga, createSchema } from 'graphql-yoga'
+import { fetch } from '@whatwg-node/fetch'
+import { GraphQLError } from 'graphql'
+
+const users = [
+  {
+    id: '1',
+    login: 'Laurin',
+  },
+  {
+    id: '2',
+    login: 'Saihaj',
+  },
+  {
+    id: '3',
+    login: 'Dotan',
+  },
+]
+
+// Provide your schema
+export const yoga = createYoga({
+  schema: createSchema({
+    typeDefs: /* GraphQL */ `
+      type User {
+        id: ID!
+        login: String!
+      }
+      type Query {
+        greeting: String!
+        user(byId: ID!): User!
+      }
+    `,
+    resolvers: {
+      Query: {
+        greeting: async () => {
+          // This service does not exist
+          const greeting = await fetch('http://localhost:9876/greeting').then(
+            (res) => res.text(),
+          )
+
+          return greeting
+        },
+        user: async (_, args) => {
+          const user = users.find((user) => user.id === args.byId)
+          if (!user) {
+            throw new GraphQLError(`User with id '${args.byId}' not found.`, {
+              extensions: {
+                code: 'USER_NOT_FOUND',
+                someRandomExtensions: {
+                  aaaa: 3,
+                },
+              },
+            })
+          }
+
+          return user
+        },
+      },
+    },
+  }),
+  logging: true,
+  maskedErrors: true,
+})
